Added interfaces for repositories Reorganized interfaces
