+++ Application/Application.csproj
@@ -89,10 +89,14 @@
     <Compile Include="Employees\Queries\GetEmployeesList\GetEmployeesListQueryTests.cs" />
     <Compile Include="Employees\Queries\GetEmployeesList\GetEmployeesQuery.cs" />
     <Compile Include="Employees\Queries\GetEmployeesList\IGetEmployeesListQuery.cs" />
-    <Compile Include="Interfaces\IDatabaseService.cs" />
-    <Compile Include="Interfaces\IInventoryService.cs" />
-    <Compile Include="Interfaces\IRepository.cs" />
-    <Compile Include="Interfaces\IUnitOfWork.cs" />
+    <Compile Include="Interfaces\Persistence\ICustomerRepository.cs" />
+    <Compile Include="Interfaces\Persistence\IDatabaseService.cs" />
+    <Compile Include="Interfaces\Infrastructure\IInventoryService.cs" />
+    <Compile Include="Interfaces\Persistence\IEmployeeRepository.cs" />
+    <Compile Include="Interfaces\Persistence\IProductRepository.cs" />
+    <Compile Include="Interfaces\Persistence\IRepository.cs" />
+    <Compile Include="Interfaces\Persistence\ISaleRepository.cs" />
+    <Compile Include="Interfaces\Persistence\IUnitOfWork.cs" />
     <Compile Include="Products\Queries\GetProductsList\GetProductsListQuery.cs" />
     <Compile Include="Products\Queries\GetProductsList\GetProductsListQueryTests.cs" />
     <Compile Include="Products\Queries\GetProductsList\IGetProductsListQuery.cs" />
