+++ Application/Customers/Queries/GetCustomerList/GetCustomersListQuery.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using CleanArchitecture.Application.Interfaces;
+using CleanArchitecture.Application.Interfaces.Persistence;
 
 namespace CleanArchitecture.Application.Customers.Queries.GetCustomerList
 {
