+++ Application/Employees/Queries/GetEmployeesList/GetEmployeesQuery.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using CleanArchitecture.Application.Interfaces;
+using CleanArchitecture.Application.Interfaces.Persistence;
 
 namespace CleanArchitecture.Application.Employees.Queries.GetEmployeesList
 {
