--- Application/Interfaces/IRepository.cs
@@ -1,20 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Security.Cryptography.X509Certificates;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace CleanArchitecture.Application.Interfaces
-{
-    public interface IRepository<T>
-    {
-        IQueryable<T> GetAll();
-
-        T Get(int id);
-
-        void Add(T entity);
-
-        void Remove(T entity);
-    }
-}
