+++ Application/Interfaces/Infrastructure/IInventoryService.cs
@@ -1,0 +1,7 @@
+﻿namespace CleanArchitecture.Application.Interfaces.Infrastructure
+{
+    public interface IInventoryService
+    {
+        void NotifySaleOcurred(int productId, int quantity);
+    }
+}
