+++ Application/Interfaces/Persistence/IRepository.cs
@@ -1,0 +1,15 @@
+﻿using System.Linq;
+
+namespace CleanArchitecture.Application.Interfaces.Persistence
+{
+    public interface IRepository<T>
+    {
+        IQueryable<T> GetAll();
+
+        T Get(int id);
+
+        void Add(T entity);
+
+        void Remove(T entity);
+    }
+}
