+++ Application/Interfaces/Persistence/IUnitOfWork.cs
@@ -1,0 +1,7 @@
+﻿namespace CleanArchitecture.Application.Interfaces.Persistence
+{
+    public interface IUnitOfWork
+    {
+        void Save();
+    }
+}
