+++ Application/Sales/Commands/CreateSale/CreateSaleCommandTests.cs
@@ -5,6 +5,8 @@ using System.Linq;
 using System.Linq.Expressions;
 using AutoMoq;
 using CleanArchitecture.Application.Interfaces;
+using CleanArchitecture.Application.Interfaces.Infrastructure;
+using CleanArchitecture.Application.Interfaces.Persistence;
 using CleanArchitecture.Application.Sales.Commands.CreateSale.Factory;
 using CleanArchitecture.Common.Dates;
 using CleanArchitecture.Common.Mocks;
