+++ Application/Sales/Queries/GetSaleDetail/GetSaleDetailQuery.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using CleanArchitecture.Application.Interfaces;
+using CleanArchitecture.Application.Interfaces.Persistence;
 
 namespace CleanArchitecture.Application.Sales.Queries.GetSaleDetail
 {
