+++ Application/Sales/Queries/GetSaleDetail/GetSaleDetailQueryTests.cs
@@ -4,6 +4,7 @@ using System.Data.Entity;
 using System.Linq;
 using AutoMoq;
 using CleanArchitecture.Application.Interfaces;
+using CleanArchitecture.Application.Interfaces.Persistence;
 using CleanArchitecture.Common.Mocks;
 using CleanArchitecture.Domain.Customers;
 using CleanArchitecture.Domain.Employees;
