+++ Persistance/Sales/SaleRepositoryTests.cs
@@ -7,6 +7,7 @@ using System.Text;
 using System.Threading.Tasks;
 using AutoMoq;
 using CleanArchitecture.Application.Interfaces;
+using CleanArchitecture.Application.Interfaces.Persistence;
 using CleanArchitecture.Common.Mocks;
 using CleanArchitecture.Domain.Sales;
 using Moq;
