+++ Persistance/Shared/DatabaseService.cs
@@ -1,5 +1,6 @@
 ﻿using System.Data.Entity;
 using CleanArchitecture.Application.Interfaces;
+using CleanArchitecture.Application.Interfaces.Persistence;
 using CleanArchitecture.Domain.Common;
 using CleanArchitecture.Domain.Customers;
 using CleanArchitecture.Domain.Employees;
