+++ Persistance/Shared/UnitOfWork.cs
@@ -1,4 +1,5 @@
 ﻿using CleanArchitecture.Application.Interfaces;
+using CleanArchitecture.Application.Interfaces.Persistence;
 
 namespace CleanArchitecture.Persistance.Shared
 {
