+++ Specification/Common/AppContext.cs
@@ -3,6 +3,8 @@ using System.Collections.Generic;
 using System.Linq;
 using AutoMoq;
 using CleanArchitecture.Application.Interfaces;
+using CleanArchitecture.Application.Interfaces.Infrastructure;
+using CleanArchitecture.Application.Interfaces.Persistence;
 using CleanArchitecture.Common.Dates;
 using StructureMap;
